More sidebar stuff and rearranging
