+++ src/components/ListDelButton.js
@@ -1,0 +1,8 @@
+// import './ListDelButton.css';
+
+export default function ListDelButton(props) {
+    return (
+        <button onClick={props.deleteList()}>
+        </button>
+    )
+}
