+++ src/components/TaskDelButton.js
@@ -1,6 +1,6 @@
-import './DelButton.css';
+import './TaskDelButton.css';
 
-export default function DelButton(props) {
+export default function TaskDelButton(props) {
     return (
         <button onClick={props.deleteCompletedTasks} className='dltBtn'>
             {props.text}
